Dump of the public_html directory as of 2018-02-12
The web root (aka public_html directory) for LNL's website has become full of random garbage, but since all of the garbage predates me I do not want to remove it without backing it up in an effective manner. I have decided to version control LNL's static site, starting with this commit that contains the contents of the web root, all garbage included, as of 2018-02-12.
Yes, I mean 2018 not 2019; I am getting the data from a tarball backup I made at that time.
The only contents excluded here are the static and media directories as those are controlled by the Django site and are not static.