Dump of the public_html directory as of 2019-02-17
The web root (aka public_html directory) for LNL's website has become full of random garbage, but since all of the garbage predates me I do not want to remove it without backing it up in an effective manner. I have decided to version control LNL's static site. I started with a commit that contains the contents of the web root, all garbage included, as of 2018-02-12 because I have a backup from that date. This commit contains the contents of the web root, all garbage included, as of today 2019-02-17.
The only contents excluded here are the static and media directories because those are controlled by the Django site and the workorder directory because that is built from source that is already version-controlled.